requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor==6.7.2
django-js-asset==3.1.2
django-tinymce==4.1.0
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
